requirements.txt
PySide6[svg]
yt-dlp
requests
Pillow
pyinstaller


